+++ spec/n_plus_one_control/rspec/matchers/perform_linear_number_of_queries_spec.rb
@@ -110,5 +110,51 @@ describe NPlusOneControl::RSpec do
         expect { Post.all }.to perform_linear_number_of_queries.with_warming_up
       end
     end
+
+    context "with custom collector" do
+      before { NPlusOneControl::CollectorsRegistry.register(another_collector) }
+      after { NPlusOneControl::CollectorsRegistry.unregister(another_collector) }
+
+      let(:another_collector) do
+        NPlusOneControl::Collectors::DB.dup.tap do |collector|
+          collector.key = :secondary_db
+          collector.name = nil
+          collector.event = "sql.active_record"
+        end
+      end
+
+      context "when has linear query", :n_plus_one do
+        populate { |n| create_list(:post, n) }
+
+        specify do
+          expect { Post.find_each { |p| p.user.name } }
+            .to perform_linear_number_of_queries(slope: 1).to(:secondary_db)
+        end
+      end
+
+      context "when has linear query larger than expected slope", :n_plus_one do
+        populate { |n| create_list(:post, n) }
+
+        specify do
+          expect do
+            expect { Post.find_each { |p| "#{p.user.name} #{p.category.name}" } }
+              .to perform_linear_number_of_queries(slope: 1).to(:secondary_db)
+          end.to raise_error(RSpec::Expectations::ExpectationNotMetError, /SECONDARY_DB/)
+        end
+      end
+
+      context "when collector has name", :n_plus_one do
+        before { another_collector.name = "secondary database" }
+
+        populate { |n| create_list(:post, n) }
+
+        specify do
+          expect do
+            expect { Post.find_each { |p| "#{p.user.name} #{p.category.name}" } }
+              .to perform_linear_number_of_queries(slope: 1).to(:secondary_db)
+          end.to raise_error(RSpec::Expectations::ExpectationNotMetError, /secondary database/)
+        end
+      end
+    end
   end
 end